+++ src/main/java/com/soft1841/sm/dao/impl/TypeDAOImpl.java
@@ -33,7 +33,7 @@ public class TypeDAOImpl implements TypeDAO {
     }
 
     @Override
-    public Entity getTypeById(int id) throws SQLException {
+    public Entity getTypeById(long id) throws SQLException {
         //采用自定义带参查询语句，返回单个实体
         return Db.use().queryOne("SELECT * FROM t_type WHERE id = ? ", id);
     }
